+++ react/hooks/useDataPixel.ts
@@ -1,0 +1,31 @@
+import { useEffect, useRef } from 'react'
+import { isEmpty } from 'ramda'
+import { usePixel } from 'vtex.pixel-manager/PixelContext'
+
+const useDataPixel = (
+  data: any[] | null,
+  pageIdentifier: any,
+  isLoading = false
+) => {
+  const { push } = usePixel()
+  const previousIdRef = useRef(null)
+
+  const previousId = previousIdRef.current
+
+  useEffect(() => {
+    if (pageIdentifier && !isLoading && previousId !== pageIdentifier) {
+      if (!data || isEmpty(data)) {
+        return
+      }
+      if (Array.isArray(data)) {
+        data.forEach(push)
+      } else {
+        push(data)
+      }
+
+      previousIdRef.current = pageIdentifier
+    }
+  }, [data, isLoading, pageIdentifier, previousId, push])
+}
+
+export default useDataPixel