+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Release 1.9.0 didn't caught the changes from version 1.8.14.
+
 ## [1.9.0] - 2019-09-12 [YANKED]
 
 ### Added